started shading

+++ Documents/ENG D STUFF/VFX/RAY TRACER/light.cpp
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-Light::Light(int x, int y, int z, float I){
+Light::Light(float x, float y, float z, float I){
     light_x=x;
     light_y=y;
     light_z=z;
@@ -21,6 +21,6 @@ vector3 Light::get_position(void){
 vector3 Light::get_light_direction(vector3 point){
     vector3 V = Light::get_position();
     vector3 l = V.vec_add(V, V.vec_scal_mult(-1,point));
-            l.normalize();
-            return l;
+    l.normalize();
+    return l;
 }
